VBOX_WITH_R0_LOGGING: build fix
svn:sync-xref-src-repo-rev: r72559

+++ src/VBox/VMM/VMMR3/VMM.cpp
@@ -1,4 +1,4 @@
-/* $Id: VMM.cpp 37595 2011-06-22 18:27:14Z noreply@oracle.com $ */
+/* $Id: VMM.cpp 37699 2011-06-29 21:50:17Z knut.osmundsen@oracle.com $ */
 /** @file
  * VMM - The Virtual Machine Monitor Core.
  */
@@ -364,7 +364,7 @@ static int vmmR3InitLoggers(PVM pVM)
                 return rc;
             pVCpu->vmm.s.pR0LoggerR3->pVM        = pVM->pVMR0;
             //pVCpu->vmm.s.pR0LoggerR3->fCreated = false;
-            pVCpu->vmm.s.pR0LoggerR3->cbLogger   = cbLogger;
+            pVCpu->vmm.s.pR0LoggerR3->cbLogger   = (uint32_t)cbLogger;
             pVCpu->vmm.s.pR0LoggerR0 = MMHyperR3ToR0(pVM, pVCpu->vmm.s.pR0LoggerR3);
         }
 # endif
